Refined association between Applicants and Courses; docs included.
Assoc and dissoc operations are done on Applicants:

* applicant.add_course_attendance(course, status)
* applicant.remove_course_attendance(course)

The Attendance association stores misc. information, such as the
datetime when the association was created, the status and so on.

Accessing those is done using the backrefs, e.g.

applicant.course -- gives you all attendances
Should probably be renamed to reflect this change.

applicant.course[0].registered -- gives you the datetime of the first
course attendance, and so on.