Specify loader's subqueryload strategy; massively (!) speeds up weight calculation by keeping the number of sql queries low.
Lazy loading is the default, which forces the loader to issue separate
SQL queries for each sub-collection. Which is bad, since we go from
a single Attendance to its Applicant, in order to count the Attendances
via the backref.

With around 2500 randomly generated Attendances this took several
minutes.

Forcing eager loading brings it down to a few seconds.

Note: the weight calculation could still be done on the database side,
e.g. via func.div(1.0,_), func.count and clever joins and casts.

See:
 * http://docs.sqlalchemy.org/en/rel_0_9/orm/loading_relationships.html
 * http://docs.sqlalchemy.org/en/rel_0_9/orm/loading_relationships.html#what-kind-of-loading-to-use
 * http://docs.sqlalchemy.org/en/rel_0_9/orm/loading_relationships.html?highlight=joinedload#sqlalchemy.orm.subqueryload