global: dockerize entire application
Now that Docker finally has proper volume and network management (as of
Docker 1.10 and Compose 1.6), we can finally switch over to a fully
container-driven application. This includes the following changes:

- Moves all components into their own images. No `dist` folder anymore.

- Tunes nginx config and unifies it into a single file. Adds self-signed
  certificate for testing. This also enables HTTP/2.0.

- Enables persistence for Redis.

- Tunes uwsgi config.

- Reworks volume layout for SPZ container.

- Removes old data for setups we don't support anymore (e.g. nginx)

It requires some more testing and additions (e.g. mem+cpu limits for
containers, auto-restart) but in general I think this is a pretty
foundation for the upcoming deployment.

Signed-off-by: Marco Neumann <marco@crepererum.net>