backend: first iteration on new user model
We got request to enable actions depending on the language administraton
rights. This prepares a better ACL control system. Because we already
store user data in our data, it seems to be a clearer solution to move
away from a static configuration to a database driven solution.

There are some leftovers from the old permission system which I'm not
going to remove entirely until the next release, but that shouldn't be a
huge deal.

Signed-off-by: Marco Neumann <marco@crepererum.net>