backend: implement CSRF protection
This protects all forms and other POST requests from CSRF. This works
with multiple workers because the used session objects (which contain
the csrf tag) are stored using signed cookies:
http://flask.pocoo.org/docs/0.10/api/#flask.sessions.SecureCookieSessionInterface

Therefore, all workers can read+verify the session, get the CSRF token
and verify the token.

Please note that GET requests are NOT protected, e.g. the `logout`
route. This is considered as harmless, because these actions are not
that important. No form data is passed by GET requests, because
Flask-WTF only parses data from POST and PUT requests:
https://github.com/lepture/flask-wtf/blob/0d2992b5daaff96b1e877859db943dc6df4500f2/flask_wtf/form.py#L127

closes #21

Signed-off-by: Marco Neumann <marco@crepererum.net>