spz: hash approval information
Do not keep approval data as cleartext. We actually do not care about
this information. So we salt+hash the tag field of the approval objects.
So, only approvals for known tags can be extracted as well as
statistical information. This way, we only keep data about our primary
"customers".

Signed-off-by: Marco Neumann <marco@crepererum.net>