nginx: set CSP referrer to `origin`
The referrer value is required for CSRF protection and w/o it our
backend won't let users to submit forms. I actually wanted to set the
policy to `same-origin` but it does not seem to be implemented by any
browser. The new value `origin` only sends the domain value w/o a path
or GET parameters, but to all targets (also external domains). This is
some kind of privacy protection at least. The domain value is sufficient
for our CSRF protection system.

Signed-off-by: Marco Neumann <marco@crepererum.net>